long

+++ src/Main.java
@@ -1,7 +1,7 @@
 public class Main {
     public static void main(String[] args) {
-SalesManager salesManager = new SalesManager(new int[] {60, 9, 100, 90});
-int max = salesManager.max();
+SalesManager salesManager = new SalesManager(new long[] {60, 9, 100, 90});
+long max = salesManager.max();
         System.out.println(max);
 
     }
